move version number components to Molecules
